fix: eslintの修正

+++ admin/eslint.config.mjs
@@ -28,8 +28,8 @@ const eslintConfig = [
       "tailwindcss/no-custom-classname": "off",
       "tailwindcss/no-contradicting-classname": "error",
       "tailwindcss/no-unnecessary-arbitrary-value": "error",
-      // クォートをダブルクォートに統一
-      "quotes": ["error", "double"],
+      // クォートをシングルクォートに統一
+      "quotes": ["error", "single"],
       // 行の最大文字数を180文字に設定
       // "max-len": ["error", { "code": 180 }],
       // セミコロンを禁止
